feat: Enable server-side instrumentation via `instrument.server.ts` file
